requirements.txt
Flask==2.2.2
Jinja2==3.1.2
PyJWT==2.4.0
requests==2.28.1
SQLAlchemy==1.4.41
uWSGI==2.0.20
pytest==7.1.3